edge-case-auditor-workspace/iteration-1/python-data-pipeline/test-project/requirements.txt
pandas
psycopg2-binary
python-dotenv
schedule
requests
